Move Outcome to :core:outcome module
