+++ core/outcome/build.gradle.kts
@@ -1,0 +1,4 @@
+plugins {
+    id(ThunderbirdPlugins.Library.jvm)
+    alias(libs.plugins.android.lint)
+}
