+++ feature/funding/googleplay/build.gradle.kts
@@ -11,6 +11,7 @@ dependencies {
     api(projects.feature.funding.api)
 
     implementation(projects.core.common)
+    implementation(projects.core.outcome)
     implementation(projects.core.ui.compose.designsystem)
 
     implementation(libs.android.billing)
