+++ feature/funding/googleplay/src/main/kotlin/app/k9mail/feature/funding/googleplay/domain/BillingManager.kt
@@ -7,6 +7,9 @@ import app.k9mail.feature.funding.googleplay.domain.entity.Contribution
 import app.k9mail.feature.funding.googleplay.domain.entity.OneTimeContribution
 import app.k9mail.feature.funding.googleplay.domain.entity.RecurringContribution
 import kotlinx.coroutines.flow.StateFlow
+import net.thunderbird.core.outcome.Outcome
+import net.thunderbird.core.outcome.flatMapSuccess
+import net.thunderbird.core.outcome.mapSuccess
 
 internal class BillingManager(
     private val billingClient: DataContract.BillingClient,
