+++ settings.gradle.kts
@@ -126,6 +126,7 @@ include(
 include(
     ":core:common",
     ":core:featureflags",
+    ":core:outcome",
     ":core:testing",
     ":core:android:common",
     ":core:android:network",
